05_image_generation/pyproject.toml
[project]
name = "05-image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "google-genai>=1.36.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
